+++ src/Bundles/Storage/Shopware6/Exception/SnippetNotFoundException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace PHPUnuhi\Bundles\Storage\Shopware6\Exception;
+
+class SnippetNotFoundException extends \Exception
+{
+
+}